+++ tests/extend_test.py
@@ -83,7 +83,9 @@ class RandomTest(jtu.JaxTestCase):
     impl = self.make_custom_impl(shape=(4, 2, 7), seed=True)
     k = jax.random.key(42, impl=impl)
     self.assertEqual(k.shape, ())
-    self.assertEqual(impl, jax.random.key_impl(k))
+    self.assertEqual(k.dtype, jax.random.key_dtype(impl))
+    impl_result = jax.random.key_impl(k)
+    self.assertEqual(impl, impl_result)
 
   def test_key_wrap_with_custom_impl(self):
     shape = (4, 2, 7)
@@ -91,13 +93,32 @@ class RandomTest(jtu.JaxTestCase):
     data = jnp.ones((3, *shape), dtype=jnp.dtype('uint32'))
     k = jax.random.wrap_key_data(data, impl=impl)
     self.assertEqual(k.shape, (3,))
-    self.assertEqual(impl, jax.random.key_impl(k))
+    self.assertEqual(k.dtype, jax.random.key_dtype(impl))
+    impl_result = jax.random.key_impl(k)
+    self.assertEqual(impl, impl_result)
 
-  def test_key_impl_is_spec(self):
+  def test_key_make_with_custom_impl_via_dtype(self):
+    impl = self.make_custom_impl(shape=(4, 2, 7), seed=True)
+    dtype = jax.random.key_dtype(impl)
+    k = jax.random.key(42, dtype=dtype)
+    self.assertEqual(k.shape, ())
+    self.assertEqual(k.dtype, dtype)
+
+  def test_key_wrap_with_custom_impl_via_dtype(self):
+    shape = (4, 2, 7)
+    impl = self.make_custom_impl(shape=shape)
+    data = jnp.ones((3, *shape), dtype=jnp.dtype('uint32'))
+    dtype = jax.random.key_dtype(impl)
+    k = jax.random.wrap_key_data(data, dtype=dtype)
+    self.assertEqual(k.shape, (3,))
+    self.assertEqual(k.dtype, dtype)
+
+  def test_key_dtype_and_spec_with_custom_impl(self):
     # this is counterpart to random_test.py:
     # KeyArrayTest.test_key_impl_builtin_is_string_name
     spec_ref = self.make_custom_impl(shape=(4, 2, 7), seed=True)
     key = jax.random.key(42, impl=spec_ref)
+    self.assertEqual(key.dtype, jax.random.key_dtype(spec_ref))
     spec = jax.random.key_impl(key)
     self.assertEqual(repr(spec), f"PRNGSpec({spec_ref._impl.name!r})")